+++ src/termout.c
@@ -3370,7 +3370,7 @@ static void
         when 26:  // Keyboard Report
           child_printf("\e[?27;0;%cn", term.has_focus ? '0' : '8');
         // DEC Locator
-        when 53 case_or 55:
+        when /*53 case_or*/ 55:  // 53 was a legacy xterm mistake, dropped in 389
           child_printf("\e[?53n");
         when 56:
           child_printf("\e[?57;1n");